Refactor pcapng: move SectionHeader to its own file
